+++ src/main/java/net/geforcemods/securitycraft/api/ICodebreakable.java
@@ -58,7 +58,7 @@ public interface ICodebreakable {
 					return false;
 				}
 
-				if (codebreaker.getOrDefault(SCContent.CODEBREAKER_DATA, CodebreakerData.DEFAULT).wasRecentlyUsed())
+				if (!player.isCreative() && codebreaker.getOrDefault(SCContent.CODEBREAKER_DATA, CodebreakerData.DEFAULT).wasRecentlyUsed())
 					return false;
 
 				boolean isSuccessful = player.isCreative() || SecurityCraft.RANDOM.nextDouble() < chance;